% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmmgutcheck-package.R
\docType{package}
\name{gmmgutcheck-package}
\alias{gmmgutcheck}
\alias{gmmgutcheck-package}
\title{gmmgutcheck: Good Mythical Morning Gut Check Data}
\description{
Rhett and Link's scores from "gut check" episodes of the YouTube series Good Mythical Morning.
}
\author{
\strong{Maintainer}: Katie Allen \email{katie.larson.allen@gmail.com} (\href{https://orcid.org/0000-0002-7330-2064}{ORCID})

}
\keyword{internal}
