% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutcheck_hostess.R
\docType{data}
\name{gutcheck_hostess}
\alias{gutcheck_hostess}
\title{Good Mythical Morning Hostess cake scores}
\format{
A tibble with 23 rows and 12 variables:
\describe{
  \item{episode}{Episode name.}
  \item{date}{Date the episode appeared on YouTube.} 
  \item{food}{The type of food that was featured in the episode.}
  \item{category}{The category of flavor.}
  \item{order}{The order in which flavors were tested.}
  \item{flavor}{The flavor that was tested.}
  \item{first_to_score}{Who was the first to score the flavor?.}
  \item{rhett_score}{Rhett's score for the flavor.}
  \item{link_score}{Link's score for the flavor.}
  \item{link_scoring_chart}{Did Link use a scoring chart in this episode?.}
  \item{total_score}{The sum of Rhett's score and Link's score.}
  \item{top_4}{The final flavor ranking chosen by Rhett and Link. If NA, flavor did not make it into the top 4.}
}
}
\source{
\url{https://youtu.be/HVq-o5dL6jk}
}
\usage{
gutcheck_hostess
}
\description{
Gut check scores from the Good Mythical Morning episode "We Try EVERY Hostess Cake Flavor".
}
\keyword{datasets}
